requirements.txt
requests
pandas
python-dotenv
streamlit        # if you’re using Streamlit
flask            # if you’re using the Flask version
